requirements.txt
streamlit
openai
feedparser
requests
python-dotenv
beautifulsoup4